backened/requirements.txt
anyio==4.9.0
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
msgspec==0.19.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
wsproto==1.2.0
